requirements.txt
markdown
fastapi
google-generativeai
markdown2
jinja2
uvicorn